separate astparser and wasim
