+++ apps/test.cpp
@@ -6,27 +6,12 @@
 #include "framework/ts.h"
 #include "smt-switch/boolector_factory.h"
 #include "vexpparser/interpreter.h"
-
+#include "astparser.h"
 #include <iostream>
 #include <sstream>
 
 using namespace wasim;
 using namespace smt;
-
-//print ast tree architecture
-void traverse_and_print_ast(const verilog_expr::VExprAst::VExprAstPtr& node, int level = 0) {
-    if (!node) {
-        return;
-    }
-    std::string indent(level * 2, ' ');
-    // std::cout << indent << "Node op: " << node -> get_op()<< std::endl;
-    std::cout << indent << "Node : " << node << std::endl;
-    // std::cout << indent << "Node Verilog: " << node->to_verilog() << std::endl;
-    // traverse child ast
-    for (const auto& child : node -> get_child() ){
-        traverse_and_print_ast(child, level + 1);
-    }
-}
 
 verilog_expr::VExprAst::VExprAstPtr get_local_max_width(const verilog_expr::VExprAst::VExprAstPtr& node, wasim::SymbolicExecutor& sim, std::vector<int>& width_vec){
     
@@ -38,17 +23,6 @@ verilog_expr::VExprAst::VExprAstPtr get_local_max_width(const verilog_expr::VExp
     }
 
     switch (node->get_op()) {
-
-      // case verilog_expr::voperator::AT: {
-
-      //       auto var_sort = sim.var(node -> get_child().at(0) -> to_verilog()) -> get_sort();
-      //       int width = var_sort -> get_width();
-      //       width_vec.push_back(width);
-
-      //       new_ast = node;
-      //       break;
-      // }
-
       case verilog_expr::voperator::MK_VAR: {
 
             auto var_sort = sim.var(node -> to_verilog()) -> get_sort();
@@ -183,7 +157,6 @@ verilog_expr::VExprAst::VExprAstPtr check_ast(const verilog_expr::VExprAst::VExp
                 local_max_width = width;
               }
             }
-
             new_ast = check_ast_soft(node, sim, local_max_width);
             break;
       }
@@ -237,8 +210,6 @@ smt::Term ast2term(SmtSolver& solver, const verilog_expr::VExprAst::VExprAstPtr&
             }
             Term left = ast2term(solver, node->get_child().at(0), sim);
             Term right = ast2term(solver, node->get_child().at(1), sim);
-            // std::cout << left << std::endl;
-            // std::cout << right << std::endl;
             result = solver->make_term(Equal, left, right);
             break;
         }
@@ -250,8 +221,6 @@ smt::Term ast2term(SmtSolver& solver, const verilog_expr::VExprAst::VExprAstPtr&
             Term left = ast2term(solver, node->get_child().at(0), sim);
             Term mid = ast2term(solver, node->get_child().at(1), sim);
             Term right = ast2term(solver, node->get_child().at(2), sim);
-            // std::cout << left << std::endl;
-            // std::cout << right << std::endl;
             result = solver->make_term(Ite, left, mid, right);
             break;
         }
@@ -267,6 +236,13 @@ smt::Term ast2term(SmtSolver& solver, const verilog_expr::VExprAst::VExprAstPtr&
             Term left = ast2term(solver, node->get_child().at(0), sim);
             Term right = ast2term(solver, node->get_child().at(1), sim);
             result = solver->make_term(BVSub, left, right);
+            break;
+        }
+        // * 
+        case verilog_expr::voperator::STAR: {
+            Term left = ast2term(solver, node->get_child().at(0), sim);
+            Term right = ast2term(solver, node->get_child().at(1), sim);
+            result = solver->make_term(BVMul, left, right);
             break;
         }
         // bit and &
@@ -296,15 +272,36 @@ smt::Term ast2term(SmtSolver& solver, const verilog_expr::VExprAst::VExprAstPtr&
             result = solver->make_term(BVNot, left);
             break;
         }
+        // <<
+        case verilog_expr::voperator::LSL: {
+            Term left = ast2term(solver, node->get_child().at(0), sim);
+            Term right = ast2term(solver, node->get_child().at(1), sim);
+
+            auto left_width = (left -> get_sort()) -> get_width();
+            auto lsl_width = log2(left_width);
+            Sort lsl_sort = solver->make_sort(BV, lsl_width);
+            Term lsl_term = solver->make_term(right -> to_int(), lsl_sort);
+            result = solver->make_term(BVShl, left, lsl_term);  // [boolector] bit-width of 'e0(left)' must be a power of 2
+            break;
+        }
+        // >>
+        case verilog_expr::voperator::LSR: {
+            Term left = ast2term(solver, node->get_child().at(0), sim);
+            Term right = ast2term(solver, node->get_child().at(1), sim);
+
+            auto left_width = (left -> get_sort()) -> get_width();
+            auto lsr_width = log2(left_width);
+            Sort lsr_sort = solver->make_sort(BV, lsr_width);
+            Term lsr_term = solver->make_term(right -> to_int(), lsr_sort);
+            result = solver->make_term(BVLshr, left, lsr_term); // [boolector] bit-width of 'e0(left)' must be a power of 2
+            break;
+        }
         // AT @
         case verilog_expr::voperator::AT: {
             Term left = ast2term(solver, node->get_child().at(0), sim);
             Term right = ast2term(solver, node->get_child().at(1), sim);
-            // std::cout << "left :" << left << std::endl;
-            // std::cout << "right :" << right << std::endl;
 
             //make sure not make same symbol
-            // auto string_name = left -> to_string() + "@" + std::to_string(right -> to_int());
             auto string_name = node -> get_child().at(0) -> to_verilog() + "@" + std::to_string(right -> to_int());
             try{
                solver -> get_symbol(string_name);
@@ -312,9 +309,8 @@ smt::Term ast2term(SmtSolver& solver, const verilog_expr::VExprAst::VExprAstPtr&
             catch(const IncorrectUsageException &e){
               result = solver->make_symbol(string_name, left -> get_sort());
             }
-            // result = solver->make_symbol(left -> to_string() + "@" + std::to_string(right -> to_int()), left -> get_sort());
-            result = solver -> get_symbol(string_name);
 
+            result = solver -> get_symbol(string_name);
             break;
         }
         // concat
@@ -350,9 +346,9 @@ smt::Term ast2term(SmtSolver& solver, const verilog_expr::VExprAst::VExprAstPtr&
             }
             Sort const_int = solver->make_sort(BV, width); // there we use BV 32 substitude int, because boolector not support int type
             result = solver -> make_term(std::stoi(node -> to_verilog()), const_int);
+            // std::cout << node -> to_verilog() << "|" << width << "|" << std::stoi(node -> to_verilog()) << std::endl;
             break;
         }
-        
         
         default:
             throw std::runtime_error("Unknown operation in AST");
@@ -381,11 +377,11 @@ int main() {
   SymbolicExecutor sim(sts, solver);
   
   /*------------------------------simulation--------------------------------*/
-  // auto varmap = sim.convert( {{"rego", 0}} );
+
   sim.init();
 
     // assert parse
-  std::string my_assertion = "(out@2 == a@0 + b@1) ? 1 : 0";
+  std::string my_assertion = "out@2 == a@0 + b@1";
 
     //vexpparser (get the ast)
   Vexp::Interpreter intp;
@@ -398,12 +394,10 @@ int main() {
     std::cout << "AST constructor error:" << e.msg_ << std::endl;
   }
   auto assertion_ast = intp.GetAstRoot();
-  // traverse_and_print_ast(ass_ast);  // print ast architecture
-
 
 
     //get information from ast (get var, max cycle, term map)
-  AstParser test(assertion_ast, sim);
+  astparser::AstParser test(assertion_ast, sim);
   bool rst_en0 = 0;                             //if input signal have not rst, give 0;  if 1 -> rst = 0, else 0 -> rst = rst1(symbolic)
   test.sim_and_get_term(sim, sts, rst_en0);
   test.print_term_map();
@@ -431,7 +425,7 @@ int main() {
 
   if( (solver -> check_sat()).result == smt::SAT )
   {
-    std::cout << "assert(" + my_assertion + ") : the equation isn't always correct and has incorrect situation" << std::endl;
+    std::cout << "assert(" + my_assertion + ") : the equation isn't always correct and has counter-example" << std::endl;
   }
   else if((solver -> check_sat()).result == smt::UNSAT )
   {
