--- design/test/adder2.btor2
@@ -1,17 +1,0 @@
-; BTOR description generated by Yosys 0.44+9 (git sha1 4b9f45273, g++ 11.4.0-1ubuntu1~22.04 -fPIC -O3) for module adder.
-1 sort bitvec 4
-2 input 1 a ; adder2.v:1.37-1.38
-3 sort bitvec 3
-4 input 3 b ; adder2.v:1.51-1.52
-5 sort bitvec 1
-6 input 5 clk ; adder2.v:1.20-1.23
-7 sort bitvec 5
-8 state 7 rego
-9 output 8 out ; adder2.v:1.67-1.70
-10 state 1 rega
-11 uext 7 10 1
-12 uext 7 4 2
-13 add 7 11 12
-14 next 7 8 13
-15 next 1 10 2
-; end of yosys output
