--- design/test/simple_pipe.btor2
@@ -1,101 +1,0 @@
-; BTOR description generated by Yosys 0.17+5 (git sha1 990c9b8e1, x86_64-conda_cos6-linux-gnu-gcc 1.24.0.133_b0863d8_dirty -fvisibility-inlines-hidden -fmessage-length=0 -march=nocona -mtune=haswell -ftree-vectorize -fPIC -fstack-protector-strong -fno-plt -O2 -ffunction-sections -fdebug-prefix-map=/home/runner/work/conda-eda/conda-eda/workdir/conda-env/conda-bld/yosys_1652397199573/work=/usr/local/src/conda/yosys-0.17_7_g990c9b8e1 -fdebug-prefix-map=/data/wenjifang/anaconda3/envs/vpipe=/usr/local/src/conda-prefix -fPIC -Os -fno-merge-constants) for module simple_pipe.
-1 sort bitvec 1
-2 input 1 clk ; simple_pipe.v:1.31-1.34
-3 sort bitvec 4
-4 input 3 reg_init ; simple_pipe.v:3.22-3.30
-5 input 1 rst ; simple_pipe.v:1.47-1.50
-6 input 1 stall1in ; simple_pipe.v:2.16-2.24
-7 input 1 stall2in ; simple_pipe.v:2.37-2.45
-8 input 1 stall3in ; simple_pipe.v:2.58-2.66
-9 input 3 w ; simple_pipe.v:1.69-1.70
-10 state 3 stage3
-11 output 10 out ; simple_pipe.v:4.23-4.26
-12 const 1 0
-13 state 1 tag3
-14 init 1 13 12
-15 not 1 13
-16 state 3 reg_v_comp
-17 eq 1 16 10
-18 or 1 15 17
-19 const 1 1
-20 not 1 18
-21 and 1 19 20
-22 state 1 tag2
-23 init 1 22 12
-24 not 1 22
-25 state 3 reg_v
-26 eq 1 25 10
-27 or 1 24 26
-28 not 1 19
-29 or 1 27 28
-30 constraint 29
-31 state 3 stage1
-32 not 1 8
-33 state 1 wen_stage2
-34 ite 1 33 32 19
-35 ite 1 7 12 34
-36 state 1 wen_stage1
-37 ite 1 36 35 19
-38 ite 1 6 12 37
-39 uext 1 38 0 stage1_go ; simple_pipe.v:36.6-36.15
-40 uext 1 38 0 stage1_ready ; simple_pipe.v:35.6-35.18
-41 state 3 stage2
-42 and 1 36 35
-43 uext 1 42 0 stage2_go ; simple_pipe.v:26.6-26.15
-44 uext 1 35 0 stage2_ready ; simple_pipe.v:25.6-25.18
-45 and 1 32 33
-46 uext 1 45 0 stage3_go ; simple_pipe.v:22.6-22.15
-47 uext 1 32 0 stage3_ready ; simple_pipe.v:20.6-20.18
-48 uext 1 33 0 stage3_valid ; simple_pipe.v:21.6-21.18
-49 state 1 tag0
-50 init 1 49 19
-51 state 1 tag1
-52 init 1 51 12
-53 ite 3 33 41 10
-54 ite 3 45 53 10
-55 ite 3 5 10 54
-56 next 3 10 55
-57 ite 1 45 22 12
-58 ite 1 5 12 57
-59 next 1 13 58
-60 sort bitvec 2
-61 const 60 10
-62 uext 3 61 2
-63 mul 3 25 62
-64 uext 3 19 3
-65 add 3 63 64
-66 next 3 16 65
-67 ite 1 45 12 22
-68 ite 1 42 51 67
-69 ite 1 5 12 68
-70 next 1 22 69
-71 ite 3 5 4 25
-72 next 3 25 71
-73 uext 3 61 2
-74 mul 3 31 73
-75 uext 3 19 3
-76 add 3 74 75
-77 ite 3 36 76 53
-78 ite 3 38 77 31
-79 ite 3 5 31 78
-80 next 3 31 79
-81 ite 1 45 12 33
-82 ite 1 42 36 81
-83 ite 1 5 33 82
-84 next 1 33 83
-85 ite 1 42 12 36
-86 ite 1 38 19 85
-87 ite 1 5 36 86
-88 next 1 36 87
-89 ite 3 42 76 41
-90 ite 3 5 41 89
-91 next 3 41 90
-92 ite 1 38 12 49
-93 ite 1 5 19 92
-94 next 1 49 93
-95 ite 1 42 12 51
-96 ite 1 38 49 95
-97 ite 1 5 12 96
-98 next 1 51 97
-99 bad 21
-; end of yosys output
