+++ framework/ts.h
@@ -30,8 +30,6 @@
 #include <map>
 #include <vector>
 
-#include <regex>
-
 namespace wasim {
 
 class StateAsmpt
@@ -621,6 +619,5 @@ class TransitionSystem
   smt::TermVec propvec_;
 
 };
-
 
 }  // namespace wasim
